Revert "Big o notation"
